requirements.txt
Django==2.1.3
django-nose==1.4.6
djangorestframework==3.9.0
drftest==1.0.1
nose==1.3.7
pytz==2018.7
